backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
openai
python-dotenv
python-docx
python-pptx
firebase-admin
slowapi
pytest
httpx
pytest-asyncio